requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
langchain-experimental
tabulate
pandas
fastapi
langchain-openai